small-hotel-app: fix bug: upload image only after creating cabin or image was updated

+++ small-hotel-app/src/services/apiCabins.js
@@ -36,18 +36,20 @@ export async function createOrEditCabin(newCabin, id) {
     throw new Error('Cabins could not be created')
   }
 
-  //2. Upload image
-  const { error: storageError } = await supabase.storage
-    .from('cabin-images')
-    .upload(imageName, newCabin.image)
+  //2. Upload image if Create Cabin or image was edited
+  if (!id || !hasImagePath) {
+    const { error: storageError } = await supabase.storage
+      .from('cabin-images')
+      .upload(imageName, newCabin.image)
 
-  //3. Delete the cabin, if there was an error uploading image
-  if (storageError) {
-    await supabase.from('cabins').delete().eq('id', data.id)
-    console.error(storageError)
-    throw new Error(
-      'Cabin image could not be uploaded and the cabin was not created'
-    )
+    //3. Delete the cabin, if there was an error uploading image
+    if (storageError) {
+      await supabase.from('cabins').delete().eq('id', data.id)
+      console.error(storageError)
+      throw new Error(
+        'Cabin image could not be uploaded and the cabin was not created'
+      )
+    }
   }
 
   return data
